requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
factory-boy==3.3.0
httpx==0.25.2
pytest-cov==4.1.0
asgi-lifespan==2.1.0
greenlet==3.0.1
ruff==0.14.2
faker==19.0.0